+++ libraries/block-kit/model/src/main/kotlin/io/adventech/blockkit/model/BlockData.kt
@@ -22,18 +22,14 @@
 
 package io.adventech.blockkit.model
 
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
 data class CompletionData(
     val length: Int,
     val placeholder: String,
     val correctCompletion: String?
 )
 
-@JsonClass(generateAdapter = true)
 data class BlockData(
-    val bible: Map<String, Excerpt>?,
-    val egw: Map<String, List<AnyBlock>>?,
+    val bible: Map<String, BlockItem.Excerpt>?,
+    val egw: Map<String, List<BlockItem>>?,
     val completion: Map<String, CompletionData>?
 )
