--- libraries/block-kit/model/src/main/kotlin/io/adventech/blockkit/model/Reference.kt
@@ -1,52 +1,0 @@
-/*
- * Copyright (c) 2024. Adventech <info@adventech.io>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package io.adventech.blockkit.model
-
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-import io.adventech.blockkit.model.resource.Resource
-import io.adventech.blockkit.model.resource.ResourceDocument
-import io.adventech.blockkit.model.resource.Segment
-
-@JsonClass(generateAdapter = true)
-data class Reference(
-    override val id: String,
-    override val style: BlockStyle?,
-    override val data: BlockData?,
-    override val nested: Boolean?,
-    val segment: Segment?,
-    val target: String,
-    val scope: ReferenceScope,
-    val title: String,
-    val subtitle: String?,
-    val resource: Resource?,
-    val document: ResourceDocument?,
-) : AnyBlock
-
-@JsonClass(generateAdapter = false)
-enum class ReferenceScope {
-    @Json(name = "document") DOCUMENT,
-    @Json(name = "resource") RESOURCE,
-    @Json(name = "segment") SEGMENT,
-    UNKNOWN
-}
