+++ libraries/block-kit/model/src/main/kotlin/io/adventech/blockkit/model/resource/Segment.kt
@@ -25,7 +25,7 @@ package io.adventech.blockkit.model.resource
 import androidx.annotation.Keep
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
-import io.adventech.blockkit.model.AnyBlock
+import io.adventech.blockkit.model.BlockItem
 
 @Keep
 @JsonClass(generateAdapter = true)
@@ -41,7 +41,7 @@ data class Segment(
     val markdownSubtitle: String? = null,
     val titleBelowCover: Boolean? = null,
     val cover: String? = null,
-    val blocks: List<AnyBlock>? = null,
+    val blocks: List<BlockItem>? = null,
     val date: String? = null,
     val background: String? = null
 )
